Enhance Power Reporting + MQTT Update
Power reporting has been updated for the Hubitat Elevation Outlet device and added to the Tasmota Outlet device.
You can specify the Power Units, Power Decimal Places, Minimum reporting Level and Reporting Power Hysteresis. The Tasmota Outlet also supports accumulated totals and you can specify the Power Units and Decimal Places for this.
The Minimum Reporting Level is used to disable logging if the reported power level is below this value. The one exception is that if the previous value was above the minimum level, then it will be reported. The Hysteresis is used to prevent reporting if the updated power value doesn't exceed the hysteresis boundaries i.e. if the hysteresis is set to 6 and the previously reported power was 100 watts, then the new value will only get reported if the new value is <= 97 or >= 103. You can set optimum values to suppress excessive reporting from "chatty" devices.